# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Toei Transportation API",
        "title": "Train Status API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/fares": {
            "get": {
                "description": "Get IC card and ticket fare between two stations",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Fare"
                ],
                "summary": "Get fare",
                "parameters": [
                    {
                        "type": "string",
                        "example": "odpt.Station:Toei.Oedo.Daimon",
                        "description": "From station ID",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "odpt.Station:Toei.Asakusa.Magome",
                        "description": "To station ID",
                        "name": "to",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.Fare"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/journeys": {
            "get": {
                "description": "Search journeys between two stations. Returns the earliest journey for each number of transfers.\nStations with the same name on different lines (e.g. Shinjuku) are treated as one station.\nTimes are on the current service day (before 03:00 belongs to the previous day).",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Journey"
                ],
                "summary": "Search journeys",
                "parameters": [
                    {
                        "type": "string",
                        "example": "odpt.Station:Toei.Mita.Kasuga",
                        "description": "From station ID",
                        "name": "from",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "odpt.Station:Toei.Asakusa.Asakusa",
                        "description": "To station ID",
                        "name": "to",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "10:00",
                        "description": "Departure time (HH:MM). Defaults to now",
                        "name": "departAt",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Arrival time (HH:MM). Cannot be used with departAt",
                        "name": "arriveBy",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 3,
                        "description": "Maximum number of transfers (0-3)",
                        "name": "maxTransfers",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "example": "odpt.Railway:Toei.Oedo",
                        "description": "Comma-separated railway IDs to avoid",
                        "name": "avoid",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.JourneySearch"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/routes": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Railway"
                ],
                "summary": "Get railways",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.Railway"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/routes/{routeId}/stations": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Station"
                ],
                "summary": "Get stations by route",
                "parameters": [
                    {
                        "type": "string",
                        "example": "odpt.Railway:Toei.Asakusa",
                        "description": "Route ID",
                        "name": "routeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.Station"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/stations/{stationId}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Station"
                ],
                "summary": "Get station detail",
                "parameters": [
                    {
                        "type": "string",
                        "example": "odpt.Station:Toei.Asakusa.Shimbashi",
                        "description": "Station ID",
                        "name": "stationId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.StationDetail"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/status": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Status"
                ],
                "summary": "Get Train Status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/service.TrainStatus"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/trains/{trainId}/location": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Train"
                ],
                "summary": "Get train location",
                "parameters": [
                    {
                        "type": "string",
                        "example": "odpt.Train:Toei.Mita.1740T",
                        "description": "Train ID",
                        "name": "trainId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/service.TrainLocation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Health"
                ],
                "summary": "Health Check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "service.DirectionTimetable": {
            "type": "object",
            "properties": {
                "calendar": {
                    "type": "string"
                },
                "isToday": {
                    "description": "現在の運行日に適用されるダイヤか",
                    "type": "boolean"
                },
                "railDirection": {
                    "type": "string"
                },
                "timetables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Timetable"
                    }
                }
            }
        },
        "service.Fare": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "icFare": {
                    "type": "integer"
                },
                "ticketFare": {
                    "type": "integer"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "service.Journey": {
            "type": "object",
            "properties": {
                "arrivalTime": {
                    "type": "string"
                },
                "departureTime": {
                    "type": "string"
                },
                "legs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.JourneyLeg"
                    }
                },
                "transfers": {
                    "type": "integer"
                }
            }
        },
        "service.JourneyLeg": {
            "type": "object",
            "properties": {
                "arrivalTime": {
                    "type": "string"
                },
                "departureTime": {
                    "type": "string"
                },
                "destination": {
                    "description": "大江戸線の環状部などでは行先が無く、空になる",
                    "type": "string"
                },
                "destinationName": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "fromName": {
                    "type": "string"
                },
                "railway": {
                    "type": "string"
                },
                "railwayName": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "toName": {
                    "type": "string"
                },
                "train": {
                    "type": "string"
                },
                "trainNumber": {
                    "type": "string"
                },
                "trainType": {
                    "type": "string"
                },
                "trainTypeName": {
                    "type": "string"
                }
            }
        },
        "service.JourneySearch": {
            "type": "object",
            "properties": {
                "journeys": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Journey"
                    }
                }
            }
        },
        "service.Passenger": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                }
            }
        },
        "service.Railway": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.Station": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "service.StationDetail": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "passengers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Passenger"
                    }
                },
                "timetables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.DirectionTimetable"
                    }
                },
                "trainLocationAvailable": {
                    "description": "この駅の路線で列車位置情報が配信されているか",
                    "type": "boolean"
                }
            }
        },
        "service.Timetable": {
            "type": "object",
            "properties": {
                "destination": {
                    "type": "string"
                },
                "time": {
                    "type": "string"
                },
                "trainId": {
                    "type": "string"
                },
                "trainNumber": {
                    "type": "string"
                },
                "trainType": {
                    "type": "string"
                },
                "trainTypeId": {
                    "description": "列車種別（例: 普通、急行、エアポート快特）",
                    "type": "string"
                }
            }
        },
        "service.TrainLocation": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "delay": {
                    "type": "integer"
                },
                "fromStation": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "railway": {
                    "type": "string"
                },
                "stopped": {
                    "description": "true の場合は fromStation に停車中（toStation は空）",
                    "type": "boolean"
                },
                "toStation": {
                    "type": "string"
                },
                "trainId": {
                    "type": "string"
                },
                "trainNumber": {
                    "type": "string"
                }
            }
        },
        "service.TrainStatus": {
            "type": "object",
            "properties": {
                "railway": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        }
    }
}